Layer the MoveIt2 build for Space ROS
Build the octomap directory first as it is required by the
MoveIt2 underlay and has a special installation as well as a
compile-time work-around.

Next, build the MoveIt2 underlay, then the MoveIt2 source.

Also, update a couple repo versions in the underlay repos to match
what is used in the MoveIt2 report

Signed-off-by: Michael Jeronimo <michael.jeronimo@openrobotics.org>