Add key update to workaround outdated base image.
Key rotation in the nvidia repositories has not yet been propagated to
the baseimage used by this project.
For now, update the key ourselves, we should be able to periodically
check the base image for updates that may allow us to revert this
change.

If we're "lucky" the build will fail when apt-key del has no key to
remove.