Save artifacts directly to current directory instead of subdirectory.
New tutorials will make use of these earthly targets in a workspace that
is already created rather than requiring a superdirectory containing the
earthly configuration file.

This may end up conflicting with building additional docker images with
more contents and require some finessing but it's better for the
current workflows.