Add Earthly configuration and use it to build the spaceros base image. (#15)
* Add Earthfile for building Space ROS base image.

Converting the base docker image (sans Ikos for now) to Earthly targets
was fairly straightforward. The targets are chosen someone arbitrarily
based on both expected semantics (i.e. `+build` for "build" and `+image` for "install" (by registering a Docker image).

* Add build-testing target to build and run colcon tests.

Does not yet produce test results artifact.

* Use ubuntu:focal as default base image.

The nvidia:cudagl images are required for doing some GPU work in builds
beyond the main build of Space ROS.
In the usual layered Docker approach we needed to use this as the base
base image for each build but with Earthly it is even easier to move to
a multi-stage build and we'll be able to pull in the install directory
from this Space ROS build in other later builds which will use the
cudagl base image.

* Use a remote artifact for the repos file.

There's a branch of the space-ros/space-ros repository which will
regenerate and store the ros2.repos file.
We can use directly asa remote earthly artifact.

* Don't change artifact paths.

* Switch base image to Ubuntu Jammy.

* Save artifacts directly to current directory instead of subdirectory.

New tutorials will make use of these earthly targets in a workspace that
is already created rather than requiring a superdirectory containing the
earthly configuration file.

This may end up conflicting with building additional docker images with
more contents and require some finessing but it's better for the
current workflows.

* Add trailing slash to copy to src directory.

* Add Earthly workflow for building the docker image.

* Ignore immediate src, install, and log subdirectories.

* Add job for uploading build results archive.

* Fix earthly download url.

* Add Earthfile artifact for bootstrapping.

* Fix spelling of COPY directive.

* Set up sources locally.

* Fix entrypoint name.

* Allow sources to supply output for the next step.

* Install all dependencies using apt.

They are all available on Ubuntu Jammy.

* Use the same directory structure as the existing Docker image.

* Save the install directory from a build-testing run.

* Add labels from base docker image.

* Start to add ikos support.

* Revert "Start to add ikos support."

This reverts commit 53cb43f21d9c14f5962e466b0241df770ac5e4f4.

Ikos support work will continue on the nuclearsandwich/ikos-earthly
branch.

* Update image tag.

* Support VCS_REF arg from Docker image build.

* Use the current HEAD as the default VCS ref.

* Use earthly to build the base Space ROS image.

* Omit GPU arguments and update image / repository name.

* Remove unused Dockerfile and corresponding GitHub workflow.

* Add push step when the build is on main.

* Start updating README.md for Earthly workflow.

* Clean and re-fetch sources in automatic build script.