Ensure that cobra is executed during tests (#103).
Space ROS's main Earthfile's build-testing procedure is invoked in CI to push
results of tests and static code analysis. Currently, for all packages
cobra-autosar.sarif results are empty. This is because cobra or some dependent
executables are not found when colcon test command is invoked.

This commit prefixes the call to colcon test in the Earthfile by a call to
source that brings all the necessary tools and dependencies into scope.