Pin gps_msgs version (#143).
The build process for the space_robots image is currently breaking
because a package depends on gps_msgs, but it is not being installed as
a dependency.

This commit adds that ROS 2 package explicitly to the space_robots
.repos file, making vcs to clone the repo and install the package during
compilation.

gps_msgs has not made a release since 2023, but they have included a bug
fix. This commit pins a specific commit from their repo (the current
HEAD as of the time of this writing).