Pin Navigation2 demo to humble-2024.10.0 (#197)
The current dockerfile of the navigation2 demo depends on space-ros:latest.
This may prevent the image from building if we release a new version of Space
ROS but we have not yet adapted the Navigation 2 demo as needed, breaking the
build.

This commit pins the version of the Navigation 2 demo to the latest version of
Space ROS that we have tested it with (humble-2024.10.0). We modify both the
Dockerfile and the build script so that the result is the same (and thus
unsurprising) no matter how the image is built.