Allow customizing moveit2 docker image tag (#222).
Currently, the build script for the moveit2 image picks latest as the tag,
which prevents us from easily customizing the tag during the build.

This commit adjusts the build script for the moveit2 image to pick a tag passed
in a variable as MOVEIT2_TAG. If the variable is not set, the latest is used
instead, defaulting to the current behavior.