Disable building space_robots image in Github Action (#237).
One of the the CI jobs fails to complete due to lack of space. This is
preventing us from complying with our process of always having a successful CI
build before merging PRs.

This commit disables the last stage of the docker build CI job, building the
`space_robots` image, where this problem manifests.